requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pyngrok==5.2.1
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.13
